+++ src/Database.php
@@ -601,6 +601,7 @@ class Database
         $this->migrateHostIpColumns();
         $this->ensureColumnExists('hosts', 'expires_at', 'VARCHAR(100) NULL');
         $this->ensureColumnExists('hosts', 'vip', 'TINYINT(1) NOT NULL DEFAULT 0');
+        $this->ensureColumnExists('hosts', 'lane_preference', 'VARCHAR(16) NULL');
         $this->ensureColumnExists('hosts', 'model_override', 'VARCHAR(128) NULL');
         $this->ensureColumnExists('hosts', 'reasoning_effort_override', 'VARCHAR(32) NULL');
         $this->ensureColumnExists('hosts', 'api_key_hash', 'CHAR(64) NULL');